restructure

+++ src/main/java/year2025/day8/Box.java
@@ -1,9 +1,5 @@
 package year2025.day8;
 
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.Set;
-
 class Box {
 
     long x;
@@ -14,7 +10,7 @@ class Box {
         long dx = x - other.x;
         long dy = y - other.y;
         long dz = z - other.z;
-        return (long)(dx*dx + dy*dy + dz*dz);
+        return (dx*dx + dy*dy + dz*dz);
     }
 
 }
